refactor(attributes): Fix circular depencency
