+++ lib/jsdom/living/attributes.js
@@ -1,5 +1,6 @@
 "use strict";
 const DOMException = require("domexception/webidl2js-wrapper");
+const NodeImpl = require("./nodes/Node-impl.js");
 const attrGenerated = require("./generated/Attr");
 const { asciiLowercase } = require("./helpers/strings");
 const { HTML_NS } = require("./helpers/namespaces");
@@ -24,18 +25,13 @@ exports.hasAttributeByNameNS = function (element, namespace, localName) {
   });
 };
 
-exports.changeAttribute = function (element, attribute, value) {
-  // https://dom.spec.whatwg.org/#concept-element-attributes-change
-
-  const { _localName, _namespace, _value } = attribute;
-  queueAttributeMutationRecord(element, _localName, _namespace, _value);
-
-  const oldValue = attribute._value;
-  attribute._value = value;
-
-  // Run jsdom hooks; roughly correspond to spec's "An attribute is set and an attribute is changed."
-  element._attrModified(attribute._qualifiedName, value, oldValue);
-};
+Object.defineProperty(exports, "changeAttribute", {
+  configurable: true,
+  enumerable: true,
+  get() {
+    return NodeImpl.changeAttribute;
+  }
+});
 
 exports.appendAttribute = function (element, attribute) {
   // https://dom.spec.whatwg.org/#concept-element-attributes-append
@@ -228,15 +224,6 @@ exports.setAttributeValue = function (element, localName, value, prefix, namespa
   exports.changeAttribute(element, attribute, value);
 };
 
-exports.setAnExistingAttributeValue = (attribute, value) => {
-  const element = attribute._element;
-  if (element === null) {
-    attribute._value = value;
-  } else {
-    exports.changeAttribute(element, attribute, value);
-  }
-};
-
 exports.removeAttributeByName = function (element, name) {
   // https://dom.spec.whatwg.org/#concept-element-attributes-remove-by-name
 
@@ -274,32 +261,6 @@ exports.copyAttributeList = function (sourceElement, destElement) {
 
     exports.appendAttribute(destElement, destAttr);
   }
-};
-
-exports.attributeListsEqual = function (elementA, elementB) {
-  // Needed by https://dom.spec.whatwg.org/#concept-node-equals
-
-  const listA = elementA._attributeList;
-  const listB = elementB._attributeList;
-
-  if (listA.length !== listB.length) {
-    return false;
-  }
-
-  for (let i = 0; i < listA.length; ++i) {
-    const attrA = listA[i];
-
-    if (!listB.some(attrB => equalsA(attrB))) {
-      return false;
-    }
-
-    function equalsA(attrB) {
-      return attrA._namespace === attrB._namespace && attrA._localName === attrB._localName &&
-             attrA._value === attrB._value;
-    }
-  }
-
-  return true;
 };
 
 exports.attributeNames = function (element) {
