+++ lib/jsdom/living/attributes/Attr-impl.js
@@ -1,10 +1,9 @@
 "use strict";
 
-const attributes = require("../attributes.js");
-const NodeImpl = require("../nodes/Node-impl.js").implementation;
+const NodeImpl = require("../nodes/Node-impl.js");
 const { ATTRIBUTE_NODE } = require("../node-type.js");
 
-class AttrImpl extends NodeImpl {
+class AttrImpl extends NodeImpl.implementation {
   constructor(globalObject, args, privateData) {
     super(globalObject, args, privateData);
 
@@ -42,7 +41,7 @@ class AttrImpl extends NodeImpl {
     return this._value;
   }
   set value(value) {
-    attributes.setAnExistingAttributeValue(this, value);
+    NodeImpl.setAnExistingAttributeValue(this, value);
   }
 
   get ownerElement() {
