+++ lib/jsdom/living/nodes/Node-impl.js
@@ -7,7 +7,7 @@ const { simultaneousIterators } = require("../../utils");
 const NODE_TYPE = require("../node-type");
 const NODE_DOCUMENT_POSITION = require("../node-document-position");
 const { clone, locateNamespacePrefix, locateNamespace } = require("../node");
-const attributes = require("../attributes");
+const { queueAttributeMutationRecord } = require("../helpers/mutation-observers");
 
 const NodeList = require("../generated/NodeList");
 
@@ -59,7 +59,7 @@ function nodeEquals(a, b) {
       break;
   }
 
-  if (a.nodeType === NODE_TYPE.ELEMENT_NODE && !attributes.attributeListsEqual(a, b)) {
+  if (a.nodeType === NODE_TYPE.ELEMENT_NODE && !attributeListsEqual(a, b)) {
     return false;
   }
 
@@ -467,7 +467,7 @@ class NodeImpl extends EventTargetImpl {
 
     switch (this.nodeType) {
       case NODE_TYPE.ATTRIBUTE_NODE: {
-        attributes.setAnExistingAttributeValue(this, value);
+        setAnExistingAttributeValue(this, value);
         break;
       }
       case NODE_TYPE.TEXT_NODE:
@@ -530,7 +530,7 @@ class NodeImpl extends EventTargetImpl {
       }
 
       case NODE_TYPE.ATTRIBUTE_NODE: {
-        attributes.setAnExistingAttributeValue(this, value);
+        setAnExistingAttributeValue(this, value);
         break;
       }
 
@@ -1081,7 +1081,52 @@ class NodeImpl extends EventTargetImpl {
     }
   }
 }
+exports.implementation = NodeImpl;
 
-module.exports = {
-  implementation: NodeImpl
-};
+// Needed by https://dom.spec.whatwg.org/#concept-node-equals
+function attributeListsEqual(elementA, elementB) {
+  const listA = elementA._attributeList;
+  const listB = elementB._attributeList;
+
+  if (listA.length !== listB.length) {
+    return false;
+  }
+
+  for (let i = 0; i < listA.length; ++i) {
+    const attrA = listA[i];
+
+    if (!listB.some(attrB => equalsA(attrB))) {
+      return false;
+    }
+
+    function equalsA(attrB) {
+      return attrA._namespace === attrB._namespace && attrA._localName === attrB._localName &&
+             attrA._value === attrB._value;
+    }
+  }
+
+  return true;
+}
+
+function setAnExistingAttributeValue(attribute, value) {
+  const element = attribute._element;
+  if (element === null) {
+    attribute._value = value;
+  } else {
+    changeAttribute(element, attribute, value);
+  }
+}
+exports.setAnExistingAttributeValue = setAnExistingAttributeValue;
+
+// https://dom.spec.whatwg.org/#concept-element-attributes-change
+function changeAttribute(element, attribute, value) {
+  const { _localName, _namespace, _value } = attribute;
+  queueAttributeMutationRecord(element, _localName, _namespace, _value);
+
+  const oldValue = attribute._value;
+  attribute._value = value;
+
+  // Run jsdom hooks; roughly correspond to spec's "An attribute is set and an attribute is changed."
+  element._attrModified(attribute._qualifiedName, value, oldValue);
+}
+exports.changeAttribute = changeAttribute;
